Move themes to another package
